third_party/WebKit/LayoutTests/fast/js/Promise-static-some-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS Promise.some() is fulfilled.
PASS result is undefined
PASS Promise.some(p6, p7, p8) is rejected.
PASS result.length is 3
PASS result[0] is "p6"
PASS result[1] is "p7"
PASS result[2] is "p8"
PASS Promise.some(p5, p7, p1) is fulfilled.
PASS result is "p1"
PASS Promise.some(p9) is fulfilled.
PASS result is "p2"
PASS successfullyParsed is true

TEST COMPLETE
